requirements.txt
openbb
pandas
matplotlib
numpy
PyYAML
streamlit